+++ tencentcloud/batch/v20170312/models.go
@@ -2077,6 +2077,10 @@ type Docker struct {
 
 	// 拉取Docker镜像失败时延迟时间。单位：秒。
 	DelayOnRetry *uint64 `json:"DelayOnRetry,omitempty" name:"DelayOnRetry"`
+
+	// Docker命令运行参数。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	DockerRunOption *string `json:"DockerRunOption,omitempty" name:"DockerRunOption"`
 }
 
 type EnhancedService struct {